moved nb_word to new file word_count
